fixed bug in DoubleVec (float->double)

+++ src/cpp/include/LCIOSTLTypes.h
@@ -14,7 +14,7 @@ namespace EVENT
 {
 typedef std::vector< std::string > StringVec ;
 typedef std::vector< float  > FloatVec ;
-typedef std::vector< float  > DoubleVec ;
+typedef std::vector< double  > DoubleVec ;
 typedef std::vector< int > IntVec ;
 typedef std::vector< short > ShortVec ;
